+++ hypervisor/include/arch/x86/host_pm.h
@@ -24,10 +24,17 @@ struct cpu_state_table {
 	struct cpu_state_info	state_info;
 };
 
+struct acpi_reset_reg {
+	struct acpi_generic_address reg;
+	uint8_t val;
+};
+
 struct pm_s_state_data *get_host_sstate_data(void);
 void host_enter_s3(struct pm_s_state_data *sstate_data, uint32_t pm1a_cnt_val, uint32_t pm1b_cnt_val);
 extern void asm_enter_s3(struct pm_s_state_data *sstate_data, uint32_t pm1a_cnt_val, uint32_t pm1b_cnt_val);
 extern void restore_s3_context(void);
 struct cpu_state_info *get_cpu_pm_state_info(void);
+struct acpi_reset_reg *get_host_reset_reg_data(void);
+void reset_host(void);
 
 #endif	/* HOST_PM_H */
